add mybatis-spring-boot

+++ mall-common/mall-common-util/src/main/java/com/lingyi/mall/common/util/PageParam.java
@@ -23,12 +23,12 @@ public class PageParam implements Serializable {
     private static final long serialVersionUID = 6133162532222082474L;
 
     @Schema(description = "页码", requiredMode = Schema.RequiredMode.NOT_REQUIRED, defaultValue = "1", example = "1")
-    @Min(value = 1L, message = "页码最小值有误")
-    @Max(value = Long.MAX_VALUE, message = "页码最大值有误")
-    private Long currentPage;
+    @Min(value = 1, message = "页码最小值有误")
+    @Max(value = Integer.MAX_VALUE, message = "页码最大值有误")
+    private Integer currentPage;
 
     @Schema(description = "页数", requiredMode = Schema.RequiredMode.NOT_REQUIRED, defaultValue = "10", example = "10")
-    @Min(value = 1L, message = "页数最小值有误")
-    @Max(value = Long.MAX_VALUE, message = "页数最大值有误")
-    private Long pageSize;
+    @Min(value = 1, message = "页数最小值有误")
+    @Max(value = Integer.MAX_VALUE, message = "页数最大值有误")
+    private Integer pageSize;
 }
